+++ src/Symfony/Bridge/Doctrine/DataCollector/DoctrineDataCollector.php
@@ -23,6 +23,8 @@ use Symfony\Component\HttpKernel\DataCollector\DataCollector;
  * DoctrineDataCollector.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.4
  */
 class DoctrineDataCollector extends DataCollector
 {