+++ src/Symfony/Bundle/FrameworkBundle/DataCollector/RouterDataCollector.php
@@ -19,6 +19,8 @@ use Symfony\Component\HttpKernel\DataCollector\RouterDataCollector as BaseRouter
  * RouterDataCollector.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @final since Symfony 4.4
  */
 class RouterDataCollector extends BaseRouterDataCollector
 {