+++ src/Symfony/Component/HttpKernel/DataCollector/AjaxDataCollector.php
@@ -18,6 +18,8 @@ use Symfony\Component\HttpFoundation\Response;
  * AjaxDataCollector.
  *
  * @author Bart van den Burg <bart@burgov.nl>
+ *
+ * @final since Symfony 4.4
  */
 class AjaxDataCollector extends DataCollector
 {